Externilized some messages, added new imagepopup/errormessage
Some more Strings were moved to the localization files
added a new imagepopup. This will be used when showing the attached
image of a evidence. it will take up most of the screen if the image is
big enough, which will make it easier to see the details of the image.
it's no longer possible to publish a experience if the user didn't give
a rating between one and five. a errormessage will be shown, if he tries
to.
removed some obsolete (and commented out) code